Fix: Check if column exists before removing it to not fail on new installations - thx @cschra

+++ database/migrations/2022_11_22_000100_remove_notif_mail_from_provbase.php
@@ -17,6 +17,10 @@ class RemoveNotifMailFromProvBase extends BaseMigration
      */
     public function up()
     {
+        if (! Schema::hasColumn($this->tableName, 'notif_mail')) {
+            return;
+        }
+
         Schema::table($this->tableName, function (Blueprint $table) {
             $table->dropColumn('notif_mail');
         });
